Moved all table-related APIs to `impl` package and made them package-private
