+++ client/src/main/java/io/pravega/client/tables/impl/ConditionalTableUpdateException.java
@@ -7,7 +7,7 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.client.tables;
+package io.pravega.client.tables.impl;
 
 import java.util.Collection;
 import lombok.Getter;
@@ -15,7 +15,7 @@ import lombok.Getter;
 /**
  * Exception that is thrown whenever a Conditional Update to a Table (based on versions) failed.
  */
-public class ConditionalTableUpdateException extends Exception {
+class ConditionalTableUpdateException extends Exception {
     private static final long serialVersionUID = 1L;
     /**
      * A Collection of Keys that failed conditional update validation.
